Delay instead of assertion failure on unsat core
PDR delegates the check for zero- and one-step counterexamples whenever
BMC is running in parallel, but it assumes this fact when
generalizing. This may lead to an empty blocking claus in inductive
generalization. If that happens we restart PDR to give BMC a chance to
catch up and send the disproved message to PDR.