Reverted Yices to ignore --check_sat_assume flag
Yices does not support check-sat with assumptions, so our `check-sat` is simulated with `assert+` and `retract` anyways.

We could revisit this if it turns out there is a reason why we should
use push/pop in Yices. Unsat core extraction would become more
complicated in this case, that's probably not worth it at the moment.