Use inferred types to strengthen type of stream
After type checking an equation do not discard the inferred type of the
right-hand side, instead change type of state variable on the left-hand
side if the type of the expression is stricter. Then we get a constraint
on the values of the state variable later

However, This is just a heuristic to make at least some use of the type inference
on the expression. In particular, it and will not infer types
transitively. For that we would need to re-type all expressions the
contain the variable, and then continue recursively.