Alternate syntax for merge
The SCADE6 syntax for merge is

    merge(h; A when c; B when not h);
    merge(h; (activate N every h)(i1,...,in); (activate M every not h)(j1,...,jm));

Since Kind 2 does not support clocked streams, then `when` operator can
only occur at the positions in the `merge` above. The arguments `A` and
`B` are on the base clock, so are the inputs `i1`,...,`jm` of `N` and
`M`.

We therefore support a less redundant syntax by implicitly clocking the
arguments of `merge`. The two expressions above are equivalent to

    merge(h; A; B);
    merge(h; N(i1,...,in); M(j1,...,jm));