Option --weakhcons true to enable weak hash-consing (disabled by default)
Weak hash-consing is useful to contain memory consumption of Kind 2, especially in the invariant generation processes but the performance cost is detrimental on some of our benchmarks. It could however be a useful tool for larger problems so it is now available as an option.

Parsing of command line arguments is now performed at load time of the module Flags. This means that the values of arguments can be accessed at load time of other modules.