Add support for abstract/free types
Represent abstract types as roughly equivalent to a struct with one
integer field. The integer describes a unique reference to an object of the
abstract type -- like a pointer. This field is implemented as a
different index type to regular record fields to ensure type safety.
In implementing this new index type, I have erred on the side of emulating
regular struct behaviour.

I chose struct-like this representation over the Abstr type in terms/types.ml
because it gives counterexamples, while Abstr did not when I tried it.

As an example, the test in tests/regression/falsifiable/abstract_type.lus
defines an abstract type `COUNTER`. It also defines imported (FFI) operations
for retrieving the count and incrementing the counter, which returns a
new counter with a one-higher count. For properties that fail, the
counterexamples for the abstract type look like arbitrary numbers, but
the important thing is that each number uniquely maps to an abstract
state describing the counter's internals.

For this test, I get the following counterexample. The initial counter is given a reference of 2, but the operation `get_counter` maps reference 2 to the actual count of 0. Then, incrementing counter referenced by 2 returns a new counter reference of 8. Asking this new counter for its count returns the actual incremented count of 1.

```
<Failure> Property (true -> (get_counter(c) = get_counter((pre c)))) is invalid by bounded model checking for k=1 after 0.186s.

Counterexample:
  Node test ()
    == Inputs ==
    init.COUNTER 2 2
    == Outputs ==
    c.COUNTER    2 8

  Function get_counter (test[l17c40])
    == Inputs ==
    in.COUNTER 4 2
    == Outputs ==
    out        7 0

  Function increment (test[l16c14])
    == Inputs ==
    in.COUNTER  3 2
    == Outputs ==
    out.COUNTER 6 8

  Function get_counter (test[l17c23])
    == Inputs ==
    in.COUNTER 2 8
    == Outputs ==
    out        0 1
```