Allow arrays in top-level constant declarations
There was a bug where a top-level constant declaraion like
```
const simple_array_literal : int^3 = [1, 5, 9];
```
was treated as a type error. This was happening because the type checking
for constants expected the trie of the type to be the same shape as
the trie of the values. However, for arrays, the values had a shape like
`[0]: _, [1]: _, [2]: _`
but the array in the type trie hadn't been flattened out, so it had
no nesting. I've fixed this for now by flattening out the array in the
type trie, so in this example, the original type:
```
Array Int (IntRange 0 3)
```
becomes the trie
```
[0]: Int, [1]: Int, [2]: Int
```