Fix array constructor typechecking bug
Array constructors (of the form `x^n`) were causing type errors in
arrows.

For a definition like:
```
const_zero = 0.0^5 -> pre const_zero;
```

This was being simplified to the following in LustreExpr:
```
abs_0 = 0.0;
const_zero[i0] = (abs_0 -> pre const_zero);
```

But this is a type error, because one side of the arrow is a real
and the other is an array. This commit inserts array selects to
make the types line up like so:
```
abs_0 = 0.0;
const_zero[i0] = (abs_0 -> pre const_zero[i0]);
```