Fix subrange revert to ints inside arrays
When defining an array where the expression has a subrange type, if
the bounds are not obviously satisfied then the type is reverted to
`int` and the bounds are added as properties. However, for arrays
this reversion was replacing the whole type with `int` rather than
just the subrange - so `(subrange [0,5] of int)^10` had its type
changed to just `int` rather than `int^10`.

I also added some pretty-printing functions, which I found useful
while debugging.