switch from erroring to best effort in dep ana
In dependency analysis it is assumed that every node is defined when
called. This requirement does not seem necessary and instead a phantom
node can be considered where the dependency information is supplied in a
"best effort" way. i.e. we look into the supplied expressions for any
variables that may effect the analysis, but otherwise do not consider
the node call outputs, as the node does not exist.