Create property name from expression before normalization
If a model includes an unnamed property that consists of
just a call to a node that returns a bool, Kind 2 uses
the call output as the associated state variable.
Before this change, the internal name was leaked in the output.
For an example, eee analysis output for:
tests/regression/success/const_in_contract.lus