Fix simulation of systems with imported subnodes
Kind 2 triggered a Not_found exception when simulating
systems with imported subnodes using the new front-end.
When map_top_reconstruct_and_add tried to map local
variable 'sofar' of an imported subnode without
assumptions, the code did not find the variable in
the corresponding transition system.
This was not an issue in the old front-end because
the 'sofar' variable was not being added to locals, and
the mapping was not triggered. This commit solves the issue by
preventing no-slicing from removing the sofar variable when
the contract has no assumptions.