Support AST modifications in type checking (#1246)
Refactoring to allow updates to the AST during the type checking step. Mainly, this involves updating `infer_type_expr` and `check_type_expr` to also return expressions, and lifting analogous changes across the module.

 This refactoring is necessary for PR #1242, which will modify the AST at type checking to store the inferred type args so they don't have to be re-inferred at the monomorphization step. (This PR does not implement this change yet.)

This PR should be functionally equivalent to the old head of `develop`, as currently the type checker is passing forward the same AST without modification.