Update tuple indexing syntax (#1260)
Update tuple indexing syntax from `t.%idx` to `t[idx]`, where `idx` must
be a concrete natural number in range.

The bracket syntax is now shared for tuples, maps, and arrays, so we
don't know at parse time the access type. So, we have to discriminate
these cases later in the pipeline. (We were already doing this for maps
and arrays.)

I also noticed we didn't have a user doc pages on tuples. Since we have
our own syntax that is probably not the same as Lustre V4 or V6, I added
a page.