Contract syntax 2 (#1266)
Use a dedicated block for contracts with the `con` and `noc` keywords.
With this syntax, we can revert back to using `var` and `const` (rather
than `gvar` and `gconst`), since the disambiguation is no longer
necessary.

In the future (for Lucent), we discussed the possibility of using
syntax:
```
contract begin 
   ...
end
```
But, the above would stylistically clash with the rest of the current
language, so we may roll out a bunch of syntactic changes all at once.

---------

Co-authored-by: Daniel Larraz <daniel-larraz@users.noreply.github.com>