Support sets, maps, and arrays in map value types (#1257)
The diff to `compile_map_def` (now `compile_map_or_set_def`) is smaller
than it appears. Besides the computation and use of `num_is` (which
tracks the number of indices are associated with the outermost map
type), the logic is the same. Because `num_is` is only for the outermost
map type, it means `compile_map_or_set_def` can only compile a single
index and not a list of indices (this is explained in the big comment).

It is a bit confusing; we can talk Tuesday if anything is unclear.

This PR does not yet fix another (orthogonal) bug in array compilation.
But, it was discovered in testing for this PR because it involves nested
arrays/tuples.
```
node N ()
returns (y13: [int^2, int]^3;)
let
  y13 = ['([1, 1], 1), '([1, 1], 1), '([1, 1], 1)];
  check y13[0][0][0] = 1;
tel
```